requirements.txt
PyPDF2
pytesseract
opencv-python
numpy
Pillow
transformers
torch
spacy
